eliminated for loop recursive calls still same problem
It should work. The guy before the four wall problem calls on him and when he returns false it means that all four of the previous guy's calls were false meaning he's also false. This creates a chain of calls returning false until we reach a step that had two or more valid decisions. Since that direction returned false, that steps decides to try the other decision. It sounds right, I just can't understand which part I messed up on. Instead of a for loop, I used 4 recursive calls linked by the or operator but it hasn't helped.